services/web/requirements.txt
flask>=1.0
flask_restful>=0.3.8
flask_script>=2.0.6
flask_migrate>=2.5.3
marshmallow>=3.7.1
flask_sqlalchemy>=2.4.4
flask_marshmallow>=0.13.0
marshmallow-sqlalchemy>=0.23.1
psycopg2-binary>=2.8.5
gunicorn>=20.0.4
dateutils>=2.8.1